refactor: add save computer repository save method
